# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Masha Api Documentation",
        "description": "Api docs for Masha Application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permissions",
                "description": "Retrieves all permissions available in the system.",
                "operationId": "83e09e395c82038fd34a94ef6870d5a1",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Permission ID",
                                                "type": "integer"
                                            },
                                            "code_permission": {
                                                "description": "Permission code",
                                                "type": "integer"
                                            },
                                            "permission_name": {
                                                "description": "Permission name",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "בעיה בשרת. יש לנסות שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Create a new permission",
                "description": "Creates a new permission in the system.",
                "operationId": "ba9b4376b552746e3828235ff44e9787",
                "requestBody": {
                    "description": "Permission details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission_code",
                                    "permission_name"
                                ],
                                "properties": {
                                    "permission_code": {
                                        "description": "Unique code for the permission",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "permission_name": {
                                        "description": "Name of the permission",
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permission created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "הרשאה נוצרה בהצלחה."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "קוד הרשאה כבר קיים במערכת או שם הרשאה כבר קיים במערכת."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "המשתמש אינו מורשה לבצע פעולה זו."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "התרחש בעיית שרת.יש לנסות שוב מאוחר יותר."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}